chore(ui): Update Calendar components folder structure
