+++ packages/ui/src/Calendar/components/Calendar/Calendar.module.scss
@@ -116,12 +116,3 @@
     }
   }
 }
-
-.dual-container {
-  display: flex;
-  justify-content: space-between;
-}
-.dual-container__spacer {
-  flex: none;
-  width: 30px;
-}
